src/data/requirements.txt
PyYAML==3.10
docopt==0.6.1
simplejson==3.3.1
wsgiref==0.1.2
